Don't include dddp.test as a package (moved out of dddp package).
