+++ MANIFEST.in
@@ -6,9 +6,9 @@ include requirements*.txt
 include .gitignore
 include Makefile
 exclude tox.ini
-graft dddp/test/meteor_todos
-prune dddp/test/build
-prune dddp/test/meteor_todos/.meteor/local
+graft tests/meteor_todos
+prune tests/build
+prune tests/meteor_todos/.meteor/local
 graft docs
 prune docs/_build
 prune docs/node_modules
