+++ tests/test_project/settings.py
@@ -38,7 +38,7 @@ INSTALLED_APPS = (
     'django.contrib.staticfiles',
     'dddp',
     'dddp.accounts',
-    'dddp.test.django_todos',
+    'django_todos',
 )
 
 MIDDLEWARE_CLASSES = (
@@ -51,9 +51,9 @@ MIDDLEWARE_CLASSES = (
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 )
 
-ROOT_URLCONF = 'dddp.test.test_project.urls'
+ROOT_URLCONF = 'test_project.urls'
 
-WSGI_APPLICATION = 'dddp.test.test_project.wsgi.application'
+WSGI_APPLICATION = 'test_project.wsgi.application'
 
 
 # Database
