+++ tox.ini
@@ -120,7 +120,7 @@ install_command=sh -c 'pip install -U "setuptools>=18.5" "wheel>=0.25.0" "pip>=7
 whitelist_externals=sh
 
 commands =
-    check-manifest --ignore "dddp/test/build*,dddp/test/meteor_todos/.meteor/local*"
+    check-manifest --ignore "tests/build*,tests/meteor_todos/.meteor/local*"
     {envpython} setup.py --no-user-cfg sdist --dist-dir={toxinidir}/dist/
     {envpython} setup.py --no-user-cfg bdist_wheel --dist-dir={toxinidir}/dist/
     sh -c "cd docs && sphinx-build -b html -d _build/doctrees -D latex_paper_size=a4  . _build/html"
